24. Swap Nodes in Pairs
