+++ src/test/kotlin/me/bossm0n5t3r/leetcode/swapNodesInPairs/SwapNodesInPairsTest.kt
@@ -1,12 +1,12 @@
-package swapNodesInPairs
+package me.bossm0n5t3r.leetcode.swapNodesInPairs
 
+import me.bossm0n5t3r.leetcode.utils.ListNode
+import me.bossm0n5t3r.leetcode.utils.ListNodeUtil
 import org.junit.jupiter.api.Test
-import utils.ListNode
-import utils.ListNodeUtil
 import kotlin.test.assertTrue
 
 internal class SwapNodesInPairsTest {
-    private val swapNodesInPairs = SwapNodesInPairs()
+    private val swapNodesInPairs = SwapNodesInPairs.Solution()
     private val listNodeUtil = ListNodeUtil()
 
     private data class SwapNodesInPairsTestData(
